Move delimiter to common & use it in draft input
