+++ kovtun.alexander/T2/delimiters.cpp
@@ -1,24 +1,6 @@
 #include "delimiters.hpp"
 
 
-std::istream & kovtun::operator>>(std::istream & in, kovtun::DelimiterI && delimiter)
-{
-  std::istream::sentry guard(in);
-  if (!guard)
-  {
-    return in;
-  }
-
-  char c = 0;
-  in >> c;
-  if (c != delimiter.expected)
-  {
-    in.setstate(std::ios::failbit);
-  }
-
-  return in;
-}
-
 std::istream & kovtun::operator>>(std::istream &in, kovtun::KeyI && stub)
 {
   std::istream::sentry guard(in);
