+++ kovtun.alexander/T2/delimiters.hpp
@@ -3,13 +3,10 @@
 
 #include <iostream>
 
+#include "../common/delimiter.hpp"
+
 namespace kovtun
 {
-  struct DelimiterI
-  {
-    char expected;
-  };
-
   struct KeyI
   {
     int & number;
@@ -30,7 +27,6 @@ namespace kovtun
     std::string & data;
   };
 
-  std::istream & operator>>(std::istream & in, DelimiterI && delimiter);
   std::istream & operator>>(std::istream & in, KeyI && keyStub);
   std::istream & operator>>(std::istream & in, DoubleLitI && doubleLit);
   std::istream & operator>>(std::istream & in, CharLitI && charLit);
